+++ docker/run.sh
@@ -9,7 +9,7 @@ if ! [ -f $CONFIG_PATH ]; then
   echo "! Can't find the given config files !"
 
   echo "You must mount a volume to the /var/run/epilink/config folder and put a configuration file in it."
-  echo "Then, put its path in the 'CONFIG_FILE' environment variable"
+  echo "Then, put its path in the 'CONFIG_FILE' environment variable."
 
   exit 1
 fi
@@ -19,5 +19,5 @@ if [ "$VERBOSE" = true ]; then
 fi
 
 echo
-
+# sleep 1h
 bin/epilink-backend $ARGS $CONFIG_PATH